convert to threaded + async (#1)
Moved away from a single-threaded synchronous tick-based architecture to a combination of threading (USB device access is in its own thread) and async. This should improve reliability and (in theory, I didn't test it!) reduce latency.

The sacrifice is simplicity of design, but I think the tradeoff is worth it.